+++ src/testRunner/tsconfig.json
@@ -21,6 +21,7 @@
         { "path": "../services", "prepend": true },
         { "path": "../jsTyping", "prepend": true },
         { "path": "../server", "prepend": true },
+        { "path": "../webServer", "prepend": true },
         { "path": "../typingsInstallerCore", "prepend": true },
         { "path": "../harness", "prepend": true }
     ],
@@ -205,6 +206,7 @@
         "unittests/tsserver/typingsInstaller.ts",
         "unittests/tsserver/versionCache.ts",
         "unittests/tsserver/watchEnvironment.ts",
+        "unittests/tsserver/webServer.ts",
         "unittests/debugDeprecation.ts"
     ]
 }
