+++ src/tsserver/tsconfig.json
@@ -8,6 +8,8 @@
         ]
     },
     "files": [
+        "nodeServer.ts",
+        "webServer.ts",
         "server.ts"
     ],
     "references": [
@@ -15,6 +17,7 @@
         { "path": "../services", "prepend": true },
         { "path": "../jsTyping", "prepend": true },
         { "path": "../server", "prepend": true },
+        { "path": "../webServer", "prepend": true },
         { "path": "../deprecatedCompat", "prepend": true }
     ]
 }
